ADD: Client quotes query for uploading new quotes into the database

+++ databaseOps.js
@@ -77,31 +77,40 @@ console.log(content);
 
 			for (const item of content) {
 					if (item.action === 'delete') {
+						if (item.section_id === 'p&s: illustratins' || item.section_id === 'p&s: 2d animation & motion graphics' || item.section_id === 'p&s: character design'){
 						const deleteQuery = 'DELETE FROM media_info_cards WHERE connection_id = $1'
 						const deleteValues = [item.connection_id];
-						await client.query(deleteQuery, deleteValues);		
-					}
-				if (item.section_id === 'p&s: illustrations' || item.section_id === 'p&s: 2d animation & motion graphics' || item.section_id === 'p&s: character design') {
-						const imageRes = await client.query('SELECT image_id FROM images WHERE url = $1', [item.src]);
-						const imageId = imageRes.rows[0]?.image_id;
-						const mediaInfoCardInsert = `INSERT INTO media_info_cards (connection_id, src, alt, link, image_id, section_id, title, description, subtitle, published_date) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) ON CONFLICT (connection_id) DO UPDATE SET src = EXCLUDED.src, alt = EXCLUDED.alt, link = EXCLUDED.link, image_id = EXCLUDED.image_id, section_id = EXCLUDED.section_id, title = EXCLUDED.title, description = EXCLUDED.description, subtitle = EXCLUDED.subtitle, published_date = EXCLUDED.published_date`;
-						const values = [item.connection_id, item.src, item.alt, item.link, imageId, item.section_id, item.title, item.description, item.subtitle, item.published_date];
-						await client.query(mediaInfoCardInsert, values);
+						await client.query(deleteQuery, deleteValues);
+						}
 
+						else if (item.section_id ===  'clients'){
+							const deleteQuery = 'DELETE FROM client_quotes WHERE connection_id = $1';
+							const deleteValue = [item.connection_id];
+							await client.query(deleteQuery, deleteValue);
+						}
+
+					}
+
+				if (item.section_id === 'p&s: illustrations' || item.section_id === 'p&s: 2d animation & motion graphics' || item.section_id === 'p&s: character design') {
+					const imageRes = await client.query('SELECT image_id FROM images WHERE url = $1', [item.src]);
+					const imageId = imageRes.rows[0]?.image_id;
+					const mediaInfoCardInsert = `INSERT INTO media_info_cards (connection_id, src, alt, link, image_id, section_id, title, description, subtitle, published_date) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) ON CONFLICT (connection_id) DO UPDATE SET src = EXCLUDED.src, alt = EXCLUDED.alt, link = EXCLUDED.link, image_id = EXCLUDED.image_id, section_id = EXCLUDED.section_id, title = EXCLUDED.title, description = EXCLUDED.description, subtitle = EXCLUDED.subtitle, published_date = EXCLUDED.published_date`;
+					const values = [item.connection_id, item.src, item.alt, item.link, imageId, item.section_id, item.title, item.description, item.subtitle, item.published_date];
+					await client.query(mediaInfoCardInsert, values);
 
 				} else if (item.section_id === 'illustrations') {
 						//Find image_id
-						const imageRes = await client.query('SELECT image_id FROM images WHERE url = $1', [item.src]);
-						const imageId = imageRes.rows[0]?.image_id;
-
-						const insertQuery = `INSERT INTO illustrations (connection_id, src, alt, link, image_id, section_id) VALUES ($1, $2, $3, $4, $5, $6) ON CONFLICT (connection_id) DO UPDATE SET src = EXCLUDED.src, alt = EXCLUDED.alt, link = EXCLUDED.link, image_id = EXCLUDED.image_id, section_id = EXCLUDED.section_id;`;
-				const values = [item.connection_id, item.src, item.alt, item.link, imageId, item.section_id];
-				await client.query(insertQuery, values);
+					const imageRes = await client.query('SELECT image_id FROM images WHERE url = $1', [item.src]);
+					const imageId = imageRes.rows[0]?.image_id;
+					const insertQuery = `INSERT INTO illustrations (connection_id, src, alt, link, image_id, section_id) VALUES ($1, $2, $3, $4, $5, $6) ON CONFLICT (connection_id) DO UPDATE SET src = EXCLUDED.src, alt = EXCLUDED.alt, link = EXCLUDED.link, image_id = EXCLUDED.image_id, section_id = EXCLUDED.section_id;`;
+					const values = [item.connection_id, item.src, item.alt, item.link, imageId, item.section_id];
+					await client.query(insertQuery, values);
+				} else if (item.section_id === 'clients') {
+				const insertQuery = 'INSERT INTO client_quotes (connection_id, section_id, link, description, subtitle) VALUES ($1, $2, $3, $4) ON CONFLICT (connection_id) DO UPDATE SET section_id = EXCLUDED.section_id, link = EXCLUDED.link, description = EXCLUDED.description, subtitle = EXCLUDED.subtitle';
+					const values = [item.connection_id, item.link, item.description, item.subtitle];
+					await client.query(insertQuery, values);
 					}
-
-				}
-
-
+			}
 			await client.query('COMMIT');
 
 			} catch(err){
@@ -118,7 +127,9 @@ async function getContent() {
 			const illustrationsResult = await pool.query('SELECT connection_id, src, alt, link, section_id  FROM illustrations ORDER BY connection_id');
 			 const mediaInfoCardsResult = await pool.query('SELECT connection_id, src, alt, link, section_id, title, description, subtitle, published_date FROM media_info_cards ORDER BY published_date' );
 
-			const combinedResult = [...illustrationsResult.rows, ...mediaInfoCardsResult.rows]
+			const clientQuotes = await pool.query('SELECT connection_id, section_id, link, title, description, subtitle');
+
+			const combinedResult = [...illustrationsResult.rows, ...mediaInfoCardsResult.rows, ...clientQuotes.rows];
 			return combinedResult;
 		} catch(err) {
 			console.error('Error getting illustrations', err.stack);
